Ensure compatability with pod_names=v1 after argo update
With argo 3.2, the workflow step names would be used for pod names.
However with the switch to 3.4 something changed and we had
non-lower-case step names that cannot be converted to pod names.

All step names will be lower case only going forward. But in order to avoid breaking running workflows, we need to keep the old default naming for at least one release.