Replace component-cli with ocm (#476)
* Update Makefile to run tests

Add setup-envtest to install-requirements since the etcd and
kube-apiserver binaries are required to run the tests. Add $(shell ...)
since the the expression within $(), here 'go env GOPATH' is evaluated
as a reference to a variable that likely does not exist otherwise.

* Upgrade dependencies and adjust incompatible api changes

* Implement logic with ocm library

This PR implements the logic previously based on component-cli
functionality with an implementation based on the ocm-lib. It therefore
adjust several related interfaces.
The PR also provides test for this newly implemented functionality.

* Remove legacy function and adjust interfaces

The new method requires several different parameters. It no longer needs
the OCI options, but requires to specify one or more repositories that
shall be used to resolve component references.

* Adjust test

* Cleanup and remove all remaining component-cli imports

* Revendor to remove component-cli dependency

* Downgrade argo-workflows to v3.4.13

* Update references to golang:1.20 to golang:1.21

* Add empty hack/tools/bin/.gitkeep

This bin directory is required in order to make the make format and make
check command work properly. Just like these two make-commands, the
setup itself is copied from the gardener repository.

* Replace deprecated functions

* Run make format and make check

* update to ocm v0.5.0

* re-trigger checks

* run make generate

* Fix `make generate` in pipeline step `check` (#486)

* run make generate

* Symlink repo repo root in pipeline check

Create a symlink for the repo root to /go/src/github.com/gardener when
before running `make generate` as it can only write to this path.

* add decoder to webhook to avoid nil pointer exception

* make format

* remove unnecessary code

---------

Co-authored-by: hendrikKahl <39563366+hendrikKahl@users.noreply.github.com>