Add latest-x pattern for major machine image versions (#525)
* add latest-x pattern for major machine image versions

Testrunner can parse latest-x pattern for major machine image
and fetch the correct version from a cloudprofile. It respects
expiration dates and architecture. Build versions continue to be ignored.

In case the pattern is not recognized , the version will be interpreted as
string and validation is delegated to the respective garden cluster
during shoot creation.

* expand testable cloudprofile